+++ include/geom/face.h
@@ -58,7 +58,7 @@ public:
   /**
    * \returns 2, the dimensionality of the object.
    */
-  virtual unsigned int dim () const override { return 2; }
+  virtual unsigned short dim () const override { return 2; }
 
   /**
    * \returns 0.  All 2D elements have no faces, just
